+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -46,6 +46,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `SecurityInterface` / `objc2-security-interface`.
   - `SensorKit` / `objc2-sensor-kit`.
   - `ShazamKit` / `objc2-shazam-kit`.
+  - `TVMLKit` / `objc2-tv-ml-kit`.
   - `TVServices` / `objc2-tv-services`.
   - `UserNotificationsUI` / `objc2-user-notifications-ui`.
 * Added fallback for `MTLCopyAllDevices` on non-macOS platforms.